Completed new path system for shapes.
This addes the rest of the shapes functions that return paths rather
than drawing, these are then drawn by the user with the DrawPath()
functions. This system seems to prevent the gpu stalling that was
present with the old method of creating/drawing/destroying paths every
time a shape wsa drawn.
The old system is changed slightly in that it doesn't keep creating and
destroying paths, rather it keeps one path that it updates each time a
shape is drawn. It doesn't seem to fix the stalling though.

A similar method is done for paints too.